refactor: Prepare for css-module
- Move all components into folders
