+++ src/components/IndexWideBlock/index.js
@@ -1,9 +1,9 @@
 import React          from 'react';
-import PostListConfig from '../configures/PostListConfig';
-import PostListStore  from '../stores/PostListStore';
-import RouteAction    from '../actions/RouteAction';
-import PostListAction from '../actions/PostListAction';
-import Elevator       from '../vendors/elevator';
+import PostListConfig from '../../configures/PostListConfig';
+import PostListStore  from '../../stores/PostListStore';
+import RouteAction    from '../../actions/RouteAction';
+import PostListAction from '../../actions/PostListAction';
+import Elevator       from '../../vendors/elevator';
 
 export default React.createClass({
