+++ src/pages/ViewPostListAuthorPage/index.js
@@ -1,10 +1,10 @@
 import React         from 'react';
-import IndexMenu     from '../components/IndexMenu';
-import CategoryMenu  from '../components/CategoryMenu';
-import IndexSection  from '../components/IndexSection';
-import Footer        from '../components/Footer';
-import RouteStore    from '../stores/RouteStore';
-import PostListStore from '../stores/PostListStore';
+import IndexMenu     from '../../components/IndexMenu';
+import CategoryMenu  from '../../components/CategoryMenu';
+import IndexSection  from '../../components/IndexSection';
+import Footer        from '../../components/Footer';
+import RouteStore    from '../../stores/RouteStore';
+import PostListStore from '../../stores/PostListStore';
 
 export default React.createClass({
 
@@ -18,7 +18,7 @@ export default React.createClass({
   },
 
   render() {
-    let tag = decodeURIComponent(RouteStore.getCurrentRoute().params.tag) || '非洲';
+    let author = decodeURIComponent(RouteStore.getCurrentRoute().params.author) || '賴奕諭';
     return (
       <div>
         <div className="fixed-top-menu">
@@ -26,7 +26,7 @@ export default React.createClass({
           <CategoryMenu />
         </div>
         <div className="container-content">
-          <IndexSection category="tag" tag={tag} title={tag}/>
+          <IndexSection category="author" author={author} title={author}/>
         </div>
         <div id="footer-divider" style={{height: '50px'}} />
         <Footer/>
